debt: Used local state for unlink confirm display
