+++ src/drive/mobile/actions/authorization.js
@@ -1,8 +1,14 @@
-import { restoreCozyClientJs, updateBarAccessToken } from '../lib/cozy-helper'
+import {
+  restoreCozyClientJs,
+  updateBarAccessToken,
+  resetClient
+} from '../lib/cozy-helper'
+import { resetPersistedState } from '../../store/persistedState'
 import { setUrl, saveCredentials } from './settings'
 
 export const REVOKE = 'REVOKE'
 export const UNREVOKE = 'UNREVOKE'
+export const UNLINK = 'UNLINK'
 
 export const revokeClient = () => ({ type: REVOKE })
 export const unrevokeClient = () => ({ type: UNREVOKE })
@@ -14,4 +20,11 @@ export const renewAuthorization = client => async dispatch => {
   updateBarAccessToken(token)
   dispatch(setUrl(url))
   dispatch(saveCredentials(infos, token))
+}
+
+export const unlink = (client, clientInfo) => async dispatch => {
+  resetClient(client, clientInfo)
+  await resetPersistedState()
+  // This action will be handled by the rootReducer: the store will be restored to its initial state
+  return dispatch({ type: UNLINK })
 }
