+++ src/drive/mobile/containers/settings/Unlink.jsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { Component } from 'react'
 import { withRouter } from 'react-router'
 import { connect } from 'react-redux'
 import { translate } from 'cozy-ui/react/I18n'
@@ -6,60 +6,62 @@ import SettingCategory, {
   ELEMENT_BUTTON
 } from '../../components/SettingCategory'
 import Modal from 'cozy-ui/react/Modal'
-import {
-  showUnlinkConfirmation,
-  hideUnlinkConfirmation,
-  unlink
-} from '../../actions/unlink'
+import { unlink } from '../../actions/authorization'
 import { getClientSettings } from '../../../reducers/settings'
 
-export const Unlink = (
-  {
-    t,
-    showUnlinkConfirmation,
-    hideUnlinkConfirmation,
-    displayUnlinkConfirmation,
-    unlink,
-    clientSettings
-  },
-  { client }
-) => (
-  <div>
-    <SettingCategory
-      title={t('mobile.settings.unlink.title')}
-      elements={[
-        {
-          type: ELEMENT_BUTTON,
-          description: t('mobile.settings.unlink.description'),
-          text: t('mobile.settings.unlink.button'),
-          theme: 'danger-outline',
-          onClick: showUnlinkConfirmation
-        }
-      ]}
-    />
-    {displayUnlinkConfirmation && (
-      <Modal
-        title={t('mobile.settings.unlink.confirmation.title')}
-        description={t('mobile.settings.unlink.confirmation.description')}
-        secondaryType="secondary"
-        secondaryText={t('mobile.settings.unlink.confirmation.cancel')}
-        secondaryAction={hideUnlinkConfirmation}
-        primaryType="danger"
-        primaryText={t('mobile.settings.unlink.confirmation.unlink')}
-        primaryAction={() => unlink(client, clientSettings)}
-      />
-    )}
-  </div>
-)
+class Unlink extends Component {
+  state = {
+    showConfirmation: false
+  }
+
+  showConfirmation = () => {
+    this.setState(state => ({ showConfirmation: true }))
+  }
+
+  hideConfirmation = () => {
+    this.setState(state => ({ showConfirmation: false }))
+  }
+
+  render() {
+    const { t, unlink, clientSettings } = this.props
+    const { showConfirmation } = this.state
+    const { client } = this.context
+    return (
+      <div>
+        <SettingCategory
+          title={t('mobile.settings.unlink.title')}
+          elements={[
+            {
+              type: ELEMENT_BUTTON,
+              description: t('mobile.settings.unlink.description'),
+              text: t('mobile.settings.unlink.button'),
+              theme: 'danger-outline',
+              onClick: this.showConfirmation
+            }
+          ]}
+        />
+        {showConfirmation && (
+          <Modal
+            title={t('mobile.settings.unlink.confirmation.title')}
+            description={t('mobile.settings.unlink.confirmation.description')}
+            secondaryType="secondary"
+            secondaryText={t('mobile.settings.unlink.confirmation.cancel')}
+            secondaryAction={this.hideConfirmation}
+            primaryType="danger"
+            primaryText={t('mobile.settings.unlink.confirmation.unlink')}
+            primaryAction={() => unlink(client, clientSettings)}
+          />
+        )}
+      </div>
+    )
+  }
+}
 
 const mapStateToProps = state => ({
-  displayUnlinkConfirmation: state.mobile.ui.displayUnlinkConfirmation,
   clientSettings: getClientSettings(state)
 })
 
 const mapDispatchToProps = (dispatch, ownProps) => ({
-  showUnlinkConfirmation: () => dispatch(showUnlinkConfirmation()),
-  hideUnlinkConfirmation: () => dispatch(hideUnlinkConfirmation()),
   unlink: (client, clientSettings) => {
     dispatch(unlink(client, clientSettings))
     ownProps.router.replace('/onboarding')
