+++ src/drive/mobile/reducers/index.js
@@ -2,12 +2,10 @@ import { combineReducers } from 'redux'
 
 import { settings } from './settings'
 import { default as mediaBackup } from '../ducks/mediaBackup'
-import { ui } from './ui'
 import { authorization } from './authorization'
 
 export default combineReducers({
   authorization,
   settings,
-  mediaBackup,
-  ui
+  mediaBackup
 })
