+++ src/drive/store/rootReducer.js
@@ -2,7 +2,7 @@
 import { combineReducers } from 'redux'
 import baseReducers from '../reducers'
 import mobileReducer from '../mobile/reducers'
-import { UNLINK } from '../mobile/actions/unlink'
+import { UNLINK } from '../mobile/actions/authorization'
 
 // Per Dan Abramov: https://stackoverflow.com/questions/35622588/how-to-reset-the-state-of-a-redux-store/35641992#35641992
 const createRootReducer = client => {
